refactor(Lint): Centralize lint files
