+++ .editorconfig
@@ -29,7 +29,7 @@ trim_trailing_whitespace = true
 # UTF-8 encoding
 charset = utf-8
 
-# Use spaces for all indents with a width of 4 characters.
+# Use spaces for all indents with a width of 2 characters.
 indent_style = space
 indent_size = 2
