+++ .github/linters/.prettierignore
@@ -1,0 +1,5 @@
+coverage/
+dist/
+storybook-static/
+tokens/
+CHANGELOG.md
