--- .prettierignore
@@ -1,3 +1,0 @@
-tokens/css/*
-tokens/dtcg/*
-CHANGELOG.md
